+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -8,5 +8,6 @@ from .gerar_arquivo_ata import gerar_arquivo_ata_async
 from .gerar_notificacao_inicio_periodo_pc import gerar_notificacao_inicio_periodo_prestacao_de_contas_async
 from .gerar_notificacao_pendencia_envio_pc import gerar_notificacao_pendencia_envio_prestacao_de_contas_async
 from .gerar_notificacao_proximidade_inicio_pc import gerar_notificacao_proximidade_inicio_periodo_prestacao_conta_async
+from .gerar_notificacao_proximidade_fim_periodo_pc import gerar_notificacao_proximidade_fim_periodo_prestacao_conta_async
