+++ sme_ptrf_apps/core/tasks/tasks_legado.py
@@ -279,20 +279,20 @@ logger = logging.getLogger(__name__)
 #     logger.info('Executado serviço de notificação de proximidade do início do período de prestação de contas async.')
 
 
-@shared_task(
-    retry_backoff=2,
-    retry_kwargs={'max_retries': 8},
-    time_limet=600,
-    soft_time_limit=300
-)
-def gerar_notificacao_proximidade_fim_periodo_prestacao_conta_async():
-    from ..core.services.notificacao_services import notificar_proximidade_fim_periodo_prestacao_conta
-
-    logger.info('Iniciando a geração de notificação de proximidade do fim do período de prestação de contas async.')
-
-    notificar_proximidade_fim_periodo_prestacao_conta()
-
-    logger.info('Finalizando a geração de notificação de proximidade do fim do período de prestação de contas async.')
+# @shared_task(
+#     retry_backoff=2,
+#     retry_kwargs={'max_retries': 8},
+#     time_limet=600,
+#     soft_time_limit=300
+# )
+# def gerar_notificacao_proximidade_fim_periodo_prestacao_conta_async():
+#     from ..core.services.notificacao_services import notificar_proximidade_fim_periodo_prestacao_conta
+#
+#     logger.info('Iniciando a geração de notificação de proximidade do fim do período de prestação de contas async.')
+#
+#     notificar_proximidade_fim_periodo_prestacao_conta()
+#
+#     logger.info('Finalizando a geração de notificação de proximidade do fim do período de prestação de contas async.')
 
 
 @shared_task(
